SSCS-10245 Delete similar object HearingLocation and refactor the code
