--- src/main/java/uk/gov/hmcts/reform/sscs/model/servicehearingvalues/HearingLocation.java
@@ -1,22 +1,0 @@
-package uk.gov.hmcts.reform.sscs.model.servicehearingvalues;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder(toBuilder = true)
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonIgnoreProperties(ignoreUnknown = true)
-@AllArgsConstructor
-@NoArgsConstructor
-public class HearingLocation {
-
-    private String locationId;
-    private String locationName;
-    private String locationType;
-    private String region;
-}
