+++ src/main/java/uk/gov/hmcts/reform/sscs/model/single/hearing/ServiceHearingValues.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sscs.model.servicehearingvalues;
+package uk.gov.hmcts.reform.sscs.model.single.hearing;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -6,6 +6,10 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import uk.gov.hmcts.reform.sscs.model.servicehearingvalues.*;
+import uk.gov.hmcts.reform.sscs.model.servicehearingvalues.HearingWindow;
+import uk.gov.hmcts.reform.sscs.model.servicehearingvalues.PanelRequirements;
+import uk.gov.hmcts.reform.sscs.model.servicehearingvalues.PartyDetails;
 
 import java.util.List;
 
@@ -28,7 +32,7 @@ public class ServiceHearingValues {
     private String hearingPriorityType;
     private Integer numberOfPhysicalAttendees;
     private boolean hearingInWelshFlag;
-    private List<HearingLocation> hearingLocations;
+    private List<HearingLocations> hearingLocations;
     private Boolean caseAdditionalSecurityFlag;
     private List<String> facilitiesRequired;
     private String listingComments;
